[LotView, VM] Terrain Flatten, Cubic Interpolation
It should probably be apparent that you should not deploy this on a
server.

+++ TSOClient/tso.world/utils/_2DSprite.cs
@@ -25,7 +25,7 @@ namespace FSO.LotView.Utils
         public Texture2D Mask;
         public Vector3 TilePosition;
         public Vector3 WorldPosition;
-        public short ObjectID; //used for mouse hit test render mode
+        public Single ObjectID; //used for mouse hit test render mode
         public ushort Room = 0xFFFF; //room to use for ambient light
         public sbyte Floor = 1;
